refactor: Double -> Float

+++ lib-semantic-api/src/commonMain/kotlin/fr/outadoc/semantique/api/model/NearbyWord.kt
@@ -3,5 +3,5 @@ package fr.outadoc.semantique.api.model
 data class NearbyWord(
     val word: String,
     val percentile: Int?,
-    val score: Double
+    val score: Float
 )
